requirements.txt
langchain
langgraph
langchain_core
langchain_google_genai
langchain_community
rich
pydantic
docker
inquirer
pyyaml